+++ integrations/github/github/core/exporters/dependabot_exporter.py
@@ -1,5 +1,6 @@
+from typing import cast
 from github.core.exporters.abstract_exporter import AbstractGithubExporter
-from github.helpers.utils import enrich_with_repository, extract_repo_params
+from github.helpers.utils import enrich_with_repository, parse_github_options
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE, RAW_ITEM
 from loguru import logger
 from github.core.options import ListDependabotAlertOptions, SingleDependabotAlertOptions
@@ -12,32 +13,34 @@ class RestDependabotAlertExporter(AbstractGithubExporter[GithubRestClient]):
         ExporterOptionsT: SingleDependabotAlertOptions
     ](self, options: ExporterOptionsT) -> RAW_ITEM:
 
-        repo_name, params = extract_repo_params(dict(options))
+        repo_name, organization, params = parse_github_options(dict(options))
         alert_number = params["alert_number"]
 
-        endpoint = f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/dependabot/alerts/{alert_number}"
+        endpoint = f"{self.client.base_url}/repos/{organization}/{repo_name}/dependabot/alerts/{alert_number}"
         response = await self.client.send_api_request(endpoint)
 
         logger.info(
-            f"Fetched Dependabot alert with number: {alert_number} for repo: {repo_name}"
+            f"Fetched Dependabot alert with number: {alert_number} for repo: {repo_name} from {organization}"
         )
 
-        return enrich_with_repository(response, repo_name)
+        return enrich_with_repository(response, cast(str, repo_name))
 
     async def get_paginated_resources[
         ExporterOptionsT: ListDependabotAlertOptions
     ](self, options: ExporterOptionsT) -> ASYNC_GENERATOR_RESYNC_TYPE:
         """Get all Dependabot alerts in the repository with pagination."""
 
-        repo_name, params = extract_repo_params(dict(options))
+        repo_name, organization, params = parse_github_options(dict(options))
         params["state"] = ",".join(params["state"])
 
         async for alerts in self.client.send_paginated_request(
-            f"{self.client.base_url}/repos/{self.client.organization}/{repo_name}/dependabot/alerts",
+            f"{self.client.base_url}/repos/{organization}/{repo_name}/dependabot/alerts",
             params,
         ):
             logger.info(
-                f"Fetched batch of {len(alerts)} Dependabot alerts from repository {repo_name}"
+                f"Fetched batch of {len(alerts)} Dependabot alerts from repository {repo_name} from {organization}"
             )
-            batch = [enrich_with_repository(alert, repo_name) for alert in alerts]
+            batch = [
+                enrich_with_repository(alert, cast(str, repo_name)) for alert in alerts
+            ]
             yield batch