+++ integrations/github/github/core/exporters/team_exporter/team_exporter.py
@@ -3,7 +3,7 @@ from loguru import logger
 
 from github.clients.http.rest_client import GithubRestClient
 from github.core.exporters.abstract_exporter import AbstractGithubExporter
-from github.core.options import SingleTeamOptions
+from github.core.options import SingleTeamOptions, ListTeamOptions
 
 
 class RestTeamExporter(AbstractGithubExporter[GithubRestClient]):
@@ -11,7 +11,7 @@ class RestTeamExporter(AbstractGithubExporter[GithubRestClient]):
         ExporterOptionT: SingleTeamOptions
     ](self, options: ExporterOptionT) -> RAW_ITEM:
         slug = options["slug"]
-        organization = self.client.organization
+        organization = options["organization"]
 
         logger.info(f"Fetching team {slug} from organization {organization}")
 
@@ -21,17 +21,25 @@ class RestTeamExporter(AbstractGithubExporter[GithubRestClient]):
         logger.info(f"Fetched team {slug} from {organization}")
         return response
 
-    async def get_paginated_resources(
-        self, options: None = None
-    ) -> ASYNC_GENERATOR_RESYNC_TYPE:
-        url = f"{self.client.base_url}/orgs/{self.client.organization}/teams"
+    async def get_paginated_resources[
+        ExporterOptionT: ListTeamOptions
+    ](self, options: ExporterOptionT) -> ASYNC_GENERATOR_RESYNC_TYPE:
+        organization = options["organization"]
+        url = f"{self.client.base_url}/orgs/{organization}/teams"
+
         async for teams in self.client.send_paginated_request(url):
+            logger.info(f"Fetched {len(teams)} teams from {organization}")
             yield teams
 
     async def get_team_repositories_by_slug[
         ExporterOptionT: SingleTeamOptions
     ](self, options: ExporterOptionT) -> ASYNC_GENERATOR_RESYNC_TYPE:
-        url = f"{self.client.base_url}/orgs/{self.client.organization}/teams/{options['slug']}/repos"
+        organization = options["organization"]
+        url = (
+            f"{self.client.base_url}/orgs/{organization}/teams/{options['slug']}/repos"
+        )
         async for repos in self.client.send_paginated_request(url):
-            logger.info(f"Fetched {len(repos)} repos for team {options['slug']}")
+            logger.info(
+                f"Fetched {len(repos)} repos for team {options['slug']} from {organization}"
+            )
             yield repos