+++ integrations/github/github/webhook/webhook_processors/check_runs/check_runs_validator_webhook_processor.py
@@ -46,28 +46,30 @@ class CheckRunValidatorWebhookProcessor(PullRequestWebhookProcessor):
         repo_name = payload["repository"]["name"]
         base_sha = pull_request["base"]["sha"]
         head_sha = pull_request["head"]["sha"]
+        organization = payload["organization"]["login"]
 
         if action not in ["opened", "synchronize", "reopened", "edited"]:
             logger.info(
-                f"Skipping handling of file validation for pull request event: {action} for {repo_name}/{pr_number}"
+                f"Skipping handling of file validation for pull request event: {action} for {repo_name}/{pr_number} of organization: {organization}"
             )
             return self._NoWebhookEventResults
 
         logger.info(
-            f"Handling file validation for pull request of type: {action} for {repo_name}/{pr_number}"
+            f"Handling file validation for pull request of type: {action} for {repo_name}/{pr_number} of organization: {organization}"
         )
 
         port_app_config = cast(GithubPortAppConfig, event.port_app_config)
-        validation_mappings = get_file_validation_mappings(port_app_config, repo_name)
+        validation_mappings = get_file_validation_mappings(port_app_config)
         repository_type = port_app_config.repository_type
 
         if not validation_mappings:
             logger.info(
-                f"No validation mappings found for repository {repo_name}, skipping validation"
+                f"No validation mappings found for repository {repo_name}, skipping validation of organization: {organization}"
             )
             return self._NoWebhookEventResults
 
         await self._handle_file_validation(
+            organization,
             repo_name,
             base_sha,
             head_sha,
@@ -80,6 +82,7 @@ class CheckRunValidatorWebhookProcessor(PullRequestWebhookProcessor):
 
     async def _handle_file_validation(
         self,
+        organization: str,
         repo_name: str,
         base_sha: str,
         head_sha: str,
@@ -88,23 +91,27 @@ class CheckRunValidatorWebhookProcessor(PullRequestWebhookProcessor):
         validation_mappings: List[ResourceConfigToPatternMapping],
     ) -> None:
         logger.info(
-            f"Fetching commit diff for repository {repo_name} from {base_sha} to {head_sha}"
+            f"Fetching commit diff for repository {repo_name} of organization: {organization} from {base_sha} to {head_sha}"
         )
 
         rest_client = create_github_client()
         file_exporter = RestFileExporter(rest_client)
-        diff_data = await file_exporter.fetch_commit_diff(repo_name, base_sha, head_sha)
+        diff_data = await file_exporter.fetch_commit_diff(
+            organization, repo_name, base_sha, head_sha
+        )
         changed_files = diff_data["files"]
 
         if not changed_files:
-            logger.debug("No changed files found, skipping validation")
+            logger.debug(
+                f"No changed files found, skipping validation of organization: {organization}"
+            )
             return
 
         logger.info(
-            f"Validation needed for {len(validation_mappings)} patterns, creating validation service"
+            f"Validation needed for {len(validation_mappings)} patterns, creating validation service of organization: {organization}"
         )
 
-        validation_service = FileValidationService()
+        validation_service = FileValidationService(organization)
 
         for validation_mapping in validation_mappings:
             files_pattern = validation_mapping.patterns