+++ integrations/github/github/webhook/webhook_processors/release_webhook_processor.py
@@ -33,9 +33,10 @@ class ReleaseWebhookProcessor(BaseRepositoryWebhookProcessor):
         repo = payload["repository"]
         release_id = release["id"]
         repo_name = repo["name"]
+        organization = payload["organization"]["login"]
 
         logger.info(
-            f"Processing release event: {action} for release {release_id} in {repo_name}"
+            f"Processing release event: {action} for release {release_id} in {repo_name} from {organization}"
         )
 
         if action in RELEASE_DELETE_EVENTS:
@@ -47,7 +48,9 @@ class ReleaseWebhookProcessor(BaseRepositoryWebhookProcessor):
         exporter = RestReleaseExporter(rest_client)
 
         data_to_upsert = await exporter.get_resource(
-            SingleReleaseOptions(repo_name=repo_name, release_id=release_id)
+            SingleReleaseOptions(
+                organization=organization, repo_name=repo_name, release_id=release_id
+            )
         )
 
         return WebhookEventRawResults(